[Toolkit] Update dependencies to add Xerces library on Windows
Tracker Reference: [Jira CORE-126](https://jira.mensiatech.com/browse/CORE-126)

+++ toolkit/src/algorithms/scenario-io/ovtkCAlgorithmScenarioImporter.cpp
@@ -305,7 +305,7 @@ boolean CAlgorithmScenarioImporter::process(void)
 		OpenViBE::Kernel::ErrorType::BadInput
 	);
 
-	if(!this->validateXML(l_pMemoryBuffer->getDirectPointer(), l_pMemoryBuffer->getSize()))
+	if(!this->validateXML(l_pMemoryBuffer->getDirectPointer(), static_cast<unsigned long>(l_pMemoryBuffer->getSize())))
 	{
 		return false; // error handling is handled in validateXML
 	}
